requirements.txt
google-generativeai>=0.3.2
python-dotenv>=1.0.0
Pillow>=10.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
